Backend/requirements.txt
# Core Python packages (required for Python 3.13)
setuptools>=65.0.0

# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database - Use psycopg (psycopg3) for Python 3.13 compatibility
psycopg[binary]>=3.1.0

# Authentication
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0

# Environment Variables
python-decouple==3.8

# ML & Analytics
textblob==0.17.1
scikit-learn>=1.4.0
pandas>=2.0.0
numpy>=1.24.0

# PDF Generation
reportlab==4.0.7
Pillow>=10.0.0

# Celery & Redis (Background Tasks) - Optional
celery==5.3.4
redis==5.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
django-extensions==3.2.3
